+++ chronograf/server/cells_test.go
@@ -13,9 +13,9 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
+	"github.com/influxdata/httprouter"
 	"github.com/influxdata/influxdb/chronograf"
 	"github.com/influxdata/influxdb/chronograf/mocks"
-	"github.com/influxdata/httprouter"
 )
 
 func Test_Cells_CorrectAxis(t *testing.T) {