+++ cmd/influx/authorization.go
@@ -279,9 +279,10 @@ var authorizationFindFlags struct {
 
 func authFindCmd() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "find",
-		Short: "Find authorization",
-		RunE:  checkSetupRunEMiddleware(&flags)(authorizationFindF),
+		Use:     "list",
+		Short:   "List authorizations",
+		Aliases: []string{"find", "ls"},
+		RunE:    checkSetupRunEMiddleware(&flags)(authorizationFindF),
 	}
 
 	cmd.Flags().StringVarP(&authorizationFindFlags.user, "user", "u", "", "The user")
@@ -309,6 +310,11 @@ func newAuthorizationService() (platform.AuthorizationService, error) {
 
 func authorizationFindF(cmd *cobra.Command, args []string) error {
 	s, err := newAuthorizationService()
+	if err != nil {
+		return err
+	}
+
+	us, err := newUserService()
 	if err != nil {
 		return err
 	}
@@ -362,11 +368,16 @@ func authorizationFindF(cmd *cobra.Command, args []string) error {
 		for _, p := range a.Permissions {
 			permissions = append(permissions, p.String())
 		}
+		user, err := us.FindUserByID(context.Background(), a.UserID)
+		if err != nil {
+			return err
+		}
 
 		w.Write(map[string]interface{}{
 			"ID":          a.ID,
 			"Token":       a.Token,
 			"Status":      a.Status,
+			"User":        user.Name,
 			"UserID":      a.UserID.String(),
 			"Permissions": permissions,
 		})
@@ -466,6 +477,11 @@ func authorizationActiveF(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
+	us, err := newUserService()
+	if err != nil {
+		return err
+	}
+
 	var id platform.ID
 	if err := id.DecodeFromString(authorizationActiveFlags.id); err != nil {
 		return err
@@ -498,10 +514,16 @@ func authorizationActiveF(cmd *cobra.Command, args []string) error {
 		ps = append(ps, p.String())
 	}
 
+	user, err := us.FindUserByID(context.Background(), a.UserID)
+	if err != nil {
+		return err
+	}
+
 	w.Write(map[string]interface{}{
 		"ID":          a.ID.String(),
 		"Token":       a.Token,
 		"Status":      a.Status,
+		"User":        user.Name,
 		"UserID":      a.UserID.String(),
 		"Permissions": ps,
 	})